requirements.txt
psycopg2
python-dotenv
django
djangorestframework